Rename "link" to "media item"
My original thought process was that users will be adding links to
media so the link was the focus. However, I'd eventually like to
be able to add media directly in some cases (e.g. photos, videos,
plain text) and maybe even use this software for physical media
(e.g. books) so it makes sense to rename this to media (item).