Make the main media link go to the view page
When viewing a list of media, my initial thought was that I want
to make it as easy as possible to get to the user's media (i.e.
opening a link) so I made the main link on the index page (i.e.
the the "title" link) go directly to the media's URL.

However, that made it annoying (at least for me) when I wanted
to get into a media item's details because generally that main
link will go to the details but not in this software. So I'd
always end up pausing over the main link and then have to go
find the "View" link which was annoying.

I think as more functionality is added, users (at least for now,
I sure do) will want to go to the media detail page more often.