Fix icons not loading in certain situations
There seems to be a bug in feather-icons that results in some
icons not being loaded correctly. For example:

<i data-feather="real-icon-1"></i>
<i data-feather="not-real"></i>
<i data-feather="real-icon-2"></i>

In the above case, the first icon will load. The second icon
causes an error because it doesn't exist, which will cause
the last icon not to load either, even though it is real.

In my case I am dynamically loading some icons based on the
media type. But due to some "legacy" data, in production there
are some media items without a type which causes an empty
data-feather. So simply removing any elements that are empty
should be fine.