Smart redirect to queue or regular media list on deletion
For example, if you go to Queues > Your queue > Your media >
Delete the media, it will redirect you to the "Your queue" page
to see the list of remaining media items. Otherwise it redirects
to the full media list page.

Previously it was always going to the media list page, which I
felt wasn't what the user would expect.