Split $() into $() for multiple and $find for single elements
I don't want to get into a bigger abstraction to allow it to
smartly know whether $('.maybe-multiple').remove() should apply
to a single element or multiple elements. So using $() to
designate that we always expect an array and $find to get a single
element should be good enough for now.

The main point was simply an alias for the longer
document.querySelectorAll anyways.